requirements.txt
requests
Faker
pyyaml
pytest